Fix circle routines All circle routines should now be the same in output; which is nice

+++ CEdev/include/lib/ce/graphx.h
@@ -261,7 +261,7 @@ void gfx_FillRectangle_NoClip(uint24_t x, uint8_t y, uint24_t width, uint8_t hei
  */
 void gfx_Circle(int x, int y, unsigned radius);
 void gfx_FillCircle(int x, int y, unsigned radius);
-void gfx_FillCircle_NoClip(uint24_t x, uint8_t y, uint8_t radius);
+void gfx_FillCircle_NoClip(uint24_t x, uint8_t y, unsigned radius);
 #define gfx_Circle_NoClip gfx_Circle
 
 /**
